requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.1
pydantic==2.9.2
pydantic-settings==2.3.4
python-multipart==0.0.9
alipay-sdk-python==3.7.796
cryptography==43.0.1
requests==2.32.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
